Add Model garden class to handle Llama2 interactions
Llama2 models hosted on VertexAI return a response object that is parsed incorrectly by the generic
VertexAIModelGarden class. Each character of the response is treated as a "prediction", and as such the
full response text is not correctly extracted.

The model response also includes the original prompt which needs to be stripted, along with the
keywords "Prompt:" and "Output:".

This Llama2 specific model introduces the following behaviors:

- Process full text returned by the model

- Extract only the newly generated text from the reponse

- Support passing in model parameter values by setting corresponding props on the object,
  this is a behavior that the core VertexAI objects support, so this adds consistency and
  good dev. "ergonmics"