Fix resolving of references in ChatVertexAI.with_structured_output
Issue #659:
When using a JSON schema with references, the method `langchain_google_vertexai.chat_models.ChatVertexAI.with_structured_output` logs a warning, stating that the key `'$defs’` is not supported. This holds if the JSON schema is passed as a dictionary, or if it is passed as a Pydantic model with the (default) keyword argument `method = None`. (The case of a Pydantic model with the keyword argument `method="json_mode"` is already handled.)

Changes:
- The implementation of the method `langchain_google_vertexai.chat_models.ChatVertexAI.with_structured_output` is changed such that in the case `method="json_mode"`, the helper function `langchain_google_vertexai._utils.replace_defs_in_schema` is called in all cases for the type of the parameter `schema` (not only if it is a Pydantic model).
- The implementation of the helper function `langchain_google_vertexai.function_utils._dict_to_gapic_schema` is changed such that an item for the key `"$defs"` in the parameter `schema` is deleted if it exists after resolving references. (Remark: This covers the case `method=None` in `langchain_google_vertexai.chat_models.ChatVertexAI.with_structured_output`.)
- The method `langchain_google_vertexai.chat_models.ChatVertexAI.with_structured_output` is changed such that a value error is raised if the parameter `schema` is neither a dictionary nor a Pydantic model.